requirements.txt
aiohttp==3.7.4
asyncio==3.4.3
nest_asyncio==1.5.1
pysmartthings==0.7
cryptography==3.4.7
psycopg2==2.8.6
python-dotenv==0.19.0
logstash
